Extract payments UI into components/helpers
Move payment-related UI and logic out of Payments.jsx into reusable modules. Adds paymentHelpers (constants, formatters, validators), PaymentBadges (Toast, Loader, EmptyState, StatusBadge, Pagination, Field, Row2), Modal, CreatePaymentModal and MarkPaidModal components, and updates Payments.jsx to import and use these new files for cleaner, modular code and easier reuse.
