+++ src/components/agent/payments/Modal.jsx
@@ -1,0 +1,28 @@
+import { useEffect } from "react";
+
+export default function Modal({ title, onClose, children, wide = false }) {
+  useEffect(() => {
+    const h = (e) => e.key === "Escape" && onClose();
+    window.addEventListener("keydown", h);
+    return () => window.removeEventListener("keydown", h);
+  }, [onClose]);
+  return (
+    <div style={{ position: "fixed", inset: 0, zIndex: 1000,
+      background: "rgba(15,23,42,0.45)",
+      display: "flex", alignItems: "center", justifyContent: "center", padding: 20 }}
+      onClick={e => e.target === e.currentTarget && onClose()}>
+      <div style={{ width: "100%", maxWidth: wide ? 680 : 500,
+        background: "#fff", borderRadius: 16,
+        boxShadow: "0 20px 60px rgba(15,23,42,0.18)",
+        maxHeight: "90vh", overflowY: "auto", animation: "fadeUp .2s ease" }}>
+        <div style={{ display: "flex", alignItems: "center", justifyContent: "space-between",
+          padding: "18px 24px", borderBottom: "1px solid #e8edf4" }}>
+          <span style={{ fontWeight: 600, fontSize: 15 }}>{title}</span>
+          <button onClick={onClose} style={{ width: 30, height: 30, border: "none",
+            background: "none", color: "#94a3b8", cursor: "pointer", fontSize: 16 }}>✕</button>
+        </div>
+        <div style={{ padding: "22px 24px" }}>{children}</div>
+      </div>
+    </div>
+  );
+}
