+++ src/components/agent/payments/PaymentBadges.jsx
@@ -1,0 +1,81 @@
+import { useEffect } from "react";
+import { STATUS_STYLE } from "./paymentHelpers";
+
+// Toast component
+export function Toast({ msg, type = "success", onDone }) {
+  useEffect(() => { const t = setTimeout(onDone, 3200); return () => clearTimeout(t); }, [onDone]);
+  return (
+    <div style={{ position: "fixed", bottom: 28, right: 28, zIndex: 9999,
+      background: type === "error" ? "#fee2e2" : "#ecfdf5",
+      color: type === "error" ? "#b91c1c" : "#047857",
+      padding: "12px 20px", borderRadius: 10, fontSize: 13.5, fontWeight: 500,
+      boxShadow: "0 4px 18px rgba(0,0,0,0.12)", maxWidth: 340 }}>{msg}</div>
+  );
+}
+
+// Loader component
+export function Loader() {
+  return (
+    <div style={{ textAlign: "center", padding: "52px 0" }}>
+      <div style={{ width: 30, height: 30, margin: "0 auto",
+        border: "3px solid #e8edf4", borderTop: "3px solid #6366f1",
+        borderRadius: "50%", animation: "spin .7s linear infinite" }} />
+    </div>
+  );
+}
+
+// EmptyState component
+export function EmptyState({ icon, text }) {
+  return (
+    <div style={{ textAlign: "center", padding: "52px 20px", color: "#94a3b8" }}>
+      <div style={{ fontSize: 36, marginBottom: 10 }}>{icon}</div>
+      <p style={{ fontSize: 14 }}>{text}</p>
+    </div>
+  );
+}
+
+// StatusBadge component
+export function StatusBadge({ status }) {
+  const s = STATUS_STYLE[status] || { bg: "#f1f5f9", color: "#64748b" };
+  return (
+    <span style={{ background: s.bg, color: s.color,
+      padding: "3px 10px", borderRadius: 20, fontSize: 11.5, fontWeight: 600 }}>
+      {status}
+    </span>
+  );
+}
+
+// Pagination component
+export function Pagination({ page, totalPages, onChange }) {
+  if (totalPages <= 1) return null;
+  return (
+    <div style={{ display: "flex", alignItems: "center", gap: 6,
+      justifyContent: "flex-end", padding: "14px 16px" }}>
+      <button className="btn btn--secondary btn--sm" disabled={page === 0}
+        onClick={() => onChange(page - 1)}>← Prev</button>
+      <span style={{ fontSize: 13, color: "#64748b", padding: "0 8px" }}>
+        {page + 1} / {totalPages}
+      </span>
+      <button className="btn btn--secondary btn--sm" disabled={page >= totalPages - 1}
+        onClick={() => onChange(page + 1)}>Next →</button>
+    </div>
+  );
+}
+
+// Field component
+export function Field({ label, children, required, hint }) {
+  return (
+    <div className="form-group">
+      <label className="form-label">
+        {label}{required && <span style={{ color: "#ef4444", marginLeft: 2 }}>*</span>}
+      </label>
+      {children}
+      {hint && <p style={{ fontSize: 11.5, color: "#94a3b8", marginTop: 4 }}>{hint}</p>}
+    </div>
+  );
+}
+
+// Row2 component (2 columns grid)
+export function Row2({ children }) {
+  return <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 14 }}>{children}</div>;
+}
