+++ src/pages/agent/Payments.jsx
@@ -2,303 +2,31 @@ import { useState, useEffect, useCallback, useRef } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
 import MainLayout from "../../components/layout/Layout";
 import api from "../../api/axios";
-
-const PAYMENT_STATUSES = ["PENDING", "PAID", "FAILED", "OVERDUE", "REFUNDED"];
-const PAYMENT_TYPES    = ["RENT", "DEPOSIT", "LATE_FEE", "MAINTENANCE"];
-const PAYMENT_METHODS  = ["BANK_TRANSFER", "CASH", "CARD", "CHECK"];
-const CURRENCIES       = ["EUR", "USD", "ALL"];
-
-const STATUS_STYLE = {
-  PENDING:  { bg: "#fffbeb", color: "#d97706" },
-  PAID:     { bg: "#ecfdf5", color: "#059669" },
-  FAILED:   { bg: "#fef2f2", color: "#dc2626" },
-  OVERDUE:  { bg: "#fff1f2", color: "#be123c" },
-  REFUNDED: { bg: "#f5f3ff", color: "#7c3aed" },
-};
-
-const fmtDate  = (d) => d ? new Date(d).toLocaleDateString("sq-AL") : "—";
-const fmtMoney = (v) => v != null ? `€${Number(v).toLocaleString("de-DE")}` : "—";
-const today    = () => new Date().toISOString().split("T")[0];
-
-function isOverdue(payment) {
-  if (!payment.due_date || payment.status !== "PENDING") return false;
-  return new Date(payment.due_date) < new Date();
-}
-
-// ─── Validim për CreatePaymentModal ──────────────────────────────────────────
-function validatePaymentForm(form, notify) {
-  if (!form.contract_id || isNaN(Number(form.contract_id)) || Number(form.contract_id) <= 0) {
-    notify("Contract ID duhet të jetë numër pozitiv", "error"); return false;
-  }
-  if (!form.amount || isNaN(Number(form.amount)) || Number(form.amount) <= 0) {
-    notify("Shuma duhet të jetë më e madhe se 0", "error"); return false;
-  }
-  if (Number(form.amount) > 999999999) {
-    notify("Shuma është shumë e madhe", "error"); return false;
-  }
-  if (!form.due_date) {
-    notify("Due date është e detyrueshme", "error"); return false;
-  }
-  if (form.notes && form.notes.length > 500) {
-    notify("Shënimet nuk mund të kalojnë 500 karaktere", "error"); return false;
-  }
-  return true;
-}
-
-// ─── Validim për MarkPaidModal ────────────────────────────────────────────────
-function validateMarkPaid(form, notify) {
-  if (!form.paid_date) {
-    notify("Data e pagesës është e detyrueshme", "error"); return false;
-  }
-  if (form.paid_date > today()) {
-    notify("Data e pagesës nuk mund të jetë në të ardhmen", "error"); return false;
-  }
-  if (form.transaction_ref && form.transaction_ref.length > 100) {
-    notify("Transaction Ref nuk mund të kalojë 100 karaktere", "error"); return false;
-  }
-  return true;
-}
-
-// ─── Shared UI ────────────────────────────────────────────────────────────────
-function Toast({ msg, type = "success", onDone }) {
-  useEffect(() => { const t = setTimeout(onDone, 3200); return () => clearTimeout(t); }, [onDone]);
-  return (
-    <div style={{ position: "fixed", bottom: 28, right: 28, zIndex: 9999,
-      background: type === "error" ? "#fee2e2" : "#ecfdf5",
-      color: type === "error" ? "#b91c1c" : "#047857",
-      padding: "12px 20px", borderRadius: 10, fontSize: 13.5, fontWeight: 500,
-      boxShadow: "0 4px 18px rgba(0,0,0,0.12)", maxWidth: 340 }}>{msg}</div>
-  );
-}
-function Loader() {
-  return (
-    <div style={{ textAlign: "center", padding: "52px 0" }}>
-      <div style={{ width: 30, height: 30, margin: "0 auto",
-        border: "3px solid #e8edf4", borderTop: "3px solid #6366f1",
-        borderRadius: "50%", animation: "spin .7s linear infinite" }} />
-    </div>
-  );
-}
-function EmptyState({ icon, text }) {
-  return (
-    <div style={{ textAlign: "center", padding: "52px 20px", color: "#94a3b8" }}>
-      <div style={{ fontSize: 36, marginBottom: 10 }}>{icon}</div>
-      <p style={{ fontSize: 14 }}>{text}</p>
-    </div>
-  );
-}
-function StatusBadge({ status }) {
-  const s = STATUS_STYLE[status] || { bg: "#f1f5f9", color: "#64748b" };
-  return (
-    <span style={{ background: s.bg, color: s.color,
-      padding: "3px 10px", borderRadius: 20, fontSize: 11.5, fontWeight: 600 }}>
-      {status}
-    </span>
-  );
-}
-function Pagination({ page, totalPages, onChange }) {
-  if (totalPages <= 1) return null;
-  return (
-    <div style={{ display: "flex", alignItems: "center", gap: 6,
-      justifyContent: "flex-end", padding: "14px 16px" }}>
-      <button className="btn btn--secondary btn--sm" disabled={page === 0}
-        onClick={() => onChange(page - 1)}>← Prev</button>
-      <span style={{ fontSize: 13, color: "#64748b", padding: "0 8px" }}>
-        {page + 1} / {totalPages}
-      </span>
-      <button className="btn btn--secondary btn--sm" disabled={page >= totalPages - 1}
-        onClick={() => onChange(page + 1)}>Next →</button>
-    </div>
-  );
-}
-function Field({ label, children, required, hint }) {
-  return (
-    <div className="form-group">
-      <label className="form-label">
-        {label}{required && <span style={{ color: "#ef4444", marginLeft: 2 }}>*</span>}
-      </label>
-      {children}
-      {hint && <p style={{ fontSize: 11.5, color: "#94a3b8", marginTop: 4 }}>{hint}</p>}
-    </div>
-  );
-}
-function Row2({ children }) {
-  return <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: 14 }}>{children}</div>;
-}
-function Modal({ title, onClose, children, wide = false }) {
-  useEffect(() => {
-    const h = (e) => e.key === "Escape" && onClose();
-    window.addEventListener("keydown", h);
-    return () => window.removeEventListener("keydown", h);
-  }, [onClose]);
-  return (
-    <div style={{ position: "fixed", inset: 0, zIndex: 1000,
-      background: "rgba(15,23,42,0.45)",
-      display: "flex", alignItems: "center", justifyContent: "center", padding: 20 }}
-      onClick={e => e.target === e.currentTarget && onClose()}>
-      <div style={{ width: "100%", maxWidth: wide ? 680 : 500,
-        background: "#fff", borderRadius: 16,
-        boxShadow: "0 20px 60px rgba(15,23,42,0.18)",
-        maxHeight: "90vh", overflowY: "auto", animation: "fadeUp .2s ease" }}>
-        <div style={{ display: "flex", alignItems: "center", justifyContent: "space-between",
-          padding: "18px 24px", borderBottom: "1px solid #e8edf4" }}>
-          <span style={{ fontWeight: 600, fontSize: 15 }}>{title}</span>
-          <button onClick={onClose} style={{ width: 30, height: 30, border: "none",
-            background: "none", color: "#94a3b8", cursor: "pointer", fontSize: 16 }}>✕</button>
-        </div>
-        <div style={{ padding: "22px 24px" }}>{children}</div>
-      </div>
-    </div>
-  );
-}
-
-// ─── Create Payment Modal ─────────────────────────────────────────────────────
-function CreatePaymentModal({ defaultContractId, onClose, onSuccess, notify }) {
-  const [form, setForm] = useState({
-    contract_id:    defaultContractId || "",
-    amount:         "",
-    currency:       "EUR",
-    payment_type:   "RENT",
-    due_date:       "",
-    payment_method: "BANK_TRANSFER",
-    notes:          "",
-  });
-  const [saving, setSaving] = useState(false);
-  const set = (k, v) => setForm(p => ({ ...p, [k]: v }));
-
-  const handleSubmit = async () => {
-    if (!validatePaymentForm(form, notify)) return;
-    setSaving(true);
-    try {
-      await api.post("/api/payments", {
-        contract_id:    Number(form.contract_id),
-        amount:         Number(form.amount),
-        currency:       form.currency,
-        payment_type:   form.payment_type,
-        due_date:       form.due_date,
-        payment_method: form.payment_method || null,
-        notes:          form.notes          || null,
-      });
-      onSuccess();
-    } catch (err) {
-      notify(err.response?.data?.message || "Gabim gjatë krijimit", "error");
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <Modal title="New Lease Payment" onClose={onClose}>
-      <Field label="Contract ID" required hint="ID numerike e kontratës">
-        <input className="form-input" type="number" min="1" value={form.contract_id}
-          onChange={e => set("contract_id", e.target.value)} placeholder="ID e kontratës" />
-      </Field>
-      <Row2>
-        <Field label="Shuma" required hint="Duhet të jetë > 0">
-          <input className="form-input" type="number" min="0.01" step="0.01" value={form.amount}
-            onChange={e => set("amount", e.target.value)} placeholder="450" />
-        </Field>
-        <Field label="Monedha">
-          <select className="form-select" value={form.currency}
-            onChange={e => set("currency", e.target.value)}>
-            {CURRENCIES.map(c => <option key={c}>{c}</option>)}
-          </select>
-        </Field>
-      </Row2>
-      <Row2>
-        <Field label="Tipi">
-          <select className="form-select" value={form.payment_type}
-            onChange={e => set("payment_type", e.target.value)}>
-            {PAYMENT_TYPES.map(t => <option key={t}>{t}</option>)}
-          </select>
-        </Field>
-        <Field label="Due Date" required>
-          <input className="form-input" type="date" value={form.due_date}
-            onChange={e => set("due_date", e.target.value)} />
-        </Field>
-      </Row2>
-      <Field label="Metoda">
-        <select className="form-select" value={form.payment_method}
-          onChange={e => set("payment_method", e.target.value)}>
-          {PAYMENT_METHODS.map(m => <option key={m}>{m}</option>)}
-        </select>
-      </Field>
-      <Field label="Shënime" hint="Max 500 karaktere">
-        <input className="form-input" value={form.notes}
-          onChange={e => set("notes", e.target.value)}
-          placeholder="(opcional)" maxLength={500} />
-      </Field>
-      <div style={{ display: "flex", gap: 10, justifyContent: "flex-end", marginTop: 6 }}>
-        <button className="btn btn--secondary" onClick={onClose}>Anulo</button>
-        <button className="btn btn--primary" onClick={handleSubmit} disabled={saving}>
-          {saving ? "Duke krijuar..." : "Krijo pagesë"}
-        </button>
-      </div>
-    </Modal>
-  );
-}
-
-// ─── Mark Paid Modal ──────────────────────────────────────────────────────────
-function MarkPaidModal({ payment, onClose, onSuccess, notify }) {
-  const [form, setForm] = useState({
-    payment_method:  payment.payment_method || "BANK_TRANSFER",
-    transaction_ref: "",
-    paid_date:       today(),
-  });
-  const [saving, setSaving] = useState(false);
-  const set = (k, v) => setForm(p => ({ ...p, [k]: v }));
-
-  const handleSubmit = async () => {
-    if (!validateMarkPaid(form, notify)) return;
-    setSaving(true);
-    try {
-      await api.patch(`/api/payments/${payment.id}/pay`, {
-        payment_method:  form.payment_method,
-        transaction_ref: form.transaction_ref || null,
-        paid_date:       form.paid_date,
-      });
-      onSuccess();
-    } catch (err) {
-      notify(err.response?.data?.message || "Gabim", "error");
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <Modal title={`Mark as PAID — Pagesë #${payment.id}`} onClose={onClose}>
-      <div style={{ background: "#ecfdf5", border: "1px solid #a7f3d0",
-        borderRadius: 8, padding: "10px 14px", marginBottom: 18, fontSize: 13, color: "#047857" }}>
-        Shuma: <strong>{fmtMoney(payment.amount)} {payment.currency}</strong>
-        {" · "}Tipi: <strong>{payment.payment_type}</strong>
-      </div>
-      <Field label="Metoda e pagesës">
-        <select className="form-select" value={form.payment_method}
-          onChange={e => set("payment_method", e.target.value)}>
-          {PAYMENT_METHODS.map(m => <option key={m}>{m}</option>)}
-        </select>
-      </Field>
-      <Row2>
-        <Field label="Transaction Ref" hint="Max 100 karaktere">
-          <input className="form-input" value={form.transaction_ref}
-            onChange={e => set("transaction_ref", e.target.value)}
-            placeholder="TXN-12345" maxLength={100} />
-        </Field>
-        <Field label="Data e pagesës" required hint="Nuk mund të jetë në të ardhmen">
-          <input className="form-input" type="date" value={form.paid_date}
-            onChange={e => set("paid_date", e.target.value)}
-            max={today()} />
-        </Field>
-      </Row2>
-      <div style={{ display: "flex", gap: 10, justifyContent: "flex-end", marginTop: 6 }}>
-        <button className="btn btn--secondary" onClick={onClose}>Anulo</button>
-        <button className="btn btn--primary" onClick={handleSubmit} disabled={saving}>
-          {saving ? "Duke shënuar..." : "✓ Konfirmo PAID"}
-        </button>
-      </div>
-    </Modal>
-  );
-}
+import {
+  PAYMENT_STATUSES,
+  PAYMENT_TYPES,
+  PAYMENT_METHODS,
+  CURRENCIES,
+  STATUS_STYLE,
+  fmtDate,
+  fmtMoney,
+  today,
+  isOverdue,
+  validatePaymentForm,
+  validateMarkPaid
+} from "../../components/agent/payments/paymentHelpers";
+import {
+  Toast,
+  Loader,
+  EmptyState,
+  StatusBadge,
+  Pagination,
+  Field,
+  Row2
+} from "../../components/agent/payments/PaymentBadges";
+import Modal from "../../components/agent/payments/Modal";
+import CreatePaymentModal from "../../components/agent/payments/CreatePaymentModal";
+import MarkPaidModal from "../../components/agent/payments/MarkPaidModal";
 
 // ═══════════════════════════════════════════════════════════════════════════════
 export default function AgentPayments() {
